# applet.benchmark: Skip going through the MODE state