@whitequark 2018-11-30
> I get the feeling that there's a reason that we're going through the MODE state but I can't figure out why.

This applet was written before applet resets worked. No reason to go through it any more.